removed console.logs and legacy TransactionPanel
