+++ src/Resources/src/components/balance/TransactionPanel.vue
@@ -1,4 +1,7 @@
 <template>
+  <BlockPopup
+    :is-open="isLoading"
+  />
   <radar-spinner
     v-if="!loaded"
     :animation-duration="2000"
@@ -13,12 +16,9 @@
       <h5 class="m-0 self-center text-[14px] text-black sm:text-[18px]">
         {{ t('balance.amount') }}: €
       </h5>
-      <input
-        ref="input"
-        type="text"
-        :value="balance < 0 ? balance.toFixed(2).slice(1).replace(/\./g, ',') : '0,00'"
-        class="h-[46px] rounded-full border-[2px] border-solid border-[#CAD6E1] bg-white text-center"
-      >
+      <MoneyInput
+        v-model="amountFieldValue"
+      />
     </div>
     <div
       id="paypal-container"
@@ -27,130 +27,114 @@
   </div>
 </template>
 
-<script>
+<script setup lang="ts">
 import { RadarSpinner } from 'epic-spinners'
 import { loadScript } from "@paypal/paypal-js";
 import { transactionStore } from "@/stores/transactionStore";
 import { useI18n } from "vue-i18n";
-import {ref} from "vue";
-import {userDataStore} from "@/stores/userDataStore";
-import {environmentStore} from "@/stores/environmentStore";
+import { WatchStopHandle, computed, onMounted, onUnmounted, ref, watch } from "vue";
+import { userDataStore } from "@/stores/userDataStore";
+import { environmentStore } from "@/stores/environmentStore";
+import postPaypalOrder from '@/api/postPaypalOrder';
+import MoneyInput from '../misc/MoneyInput.vue';
+import BlockPopup from '../misc/BlockPopup.vue';
 
-export default {
-  components: {RadarSpinner},
-  emits: ['closePanel'],
-  setup() {
-    const { t } = useI18n();
-    return { t };
-  },
-  data() {
-    return {
-      balance: userDataStore.getState().balance,
-      locale: ref('en'),
-      loaded: ref(false)
-    }
-  },
-  async mounted() {
-    let amountField = this.$refs['input'];
-    const vm = this;
+const { t } = useI18n();
 
-    const { locale } = useI18n();
+const balance = computed(() => userDataStore.getState().balance);
+const amountFieldValue = ref(0.00);
+const loaded = ref(false);
+const isLoading = ref(false);
+const actionsWatcher = ref<WatchStopHandle | null>(null);
 
-    function formatCurrency(total) {
-      let neg = false;
-      if (total < 0) {
-        neg = true;
-        total = Math.abs(total);
+const emit = defineEmits(['closePanel']);
+
+onMounted(async () => {
+  amountFieldValue.value = balance.value < 0 ? Math.abs(balance.value) : 0.00;
+  try {
+    const paypal = await loadScript({
+      'client-id': environmentStore.getState().paypalId,
+      currency: 'EUR'
+    });
+
+    paypal.Buttons({
+      onInit: function(data, actions) {
+        if (amountFieldValue.value > 0.00) {
+          actions.enable();
+        } else {
+          actions.disable();
+        }
+
+        actionsWatcher.value = watch(
+          () => amountFieldValue.value,
+          () => {
+            if (amountFieldValue.value > 0.00) {
+              actions.enable();
+            } else {
+              actions.disable();
+            }
+          }
+        );
+      },
+      createOrder: function(data, actions) {
+        return actions.order.create({
+          purchase_units: [{
+            amount: {
+              value: formatCurrency(amountFieldValue.value)
+            }
+          }]
+        });
+      },
+      onShippingChange: function(data, actions) {
+        return actions.resolve();
+      },
+      onApprove: async function (data, actions) {
+        isLoading.value = true;
+        try {
+          await actions.order.capture();
+
+          const response = await postPaypalOrder(amountFieldValue.value.toFixed(2), data.orderID);
+
+          if (response.ok) {
+            userDataStore.adjustBalance(parseFloat(formatCurrency(amountFieldValue.value)));
+            transactionStore.fillStore();
+            // disable gray out and show spinner
+            emit('closePanel');
+          } else {
+            console.log('Error from Backend');
+          }
+          isLoading.value = false;
+        } catch (error) {
+          console.log(`error on approving: ${error}`);
+          isLoading.value = false;
+        }
       }
-      return (neg ? '-' : '') + parseFloat(total.replace(/,/g, '.')).toFixed(2).toString();
-    }
-
-    loadScript({
-      "client-id": environmentStore.getState().paypalId,
-      currency: "EUR",
     })
-        .then((paypal) => {
-          paypal
-              .Buttons({
-                onInit: function (data, actions) {
-                  if (amountField.checkValidity() === true && parseFloat(amountField.value.replace(/,/g, '.')) > 0.00) {
-                    actions.enable();
-                  } else {
-                    actions.disable();
-                  }
+    .render('#paypal-container')
+    .catch(error => {
+      console.error('failed to render the PayPal Buttons', error);
+      isLoading.value = false;
+    });
 
-                  amountField.addEventListener('input', () => {
-                    // Replace a comma with a point and parse the input string to a float
-                    let amountFieldValue = parseFloat(amountField.value.replace(/,/g, '.'));
-
-                    amountField.setCustomValidity('');
-
-                    // If the input is valid (matches the HTML pattern: "\d*([.,]?\d{0,2})") and the value is above 0.00..
-                    if (amountField.checkValidity() === true && amountFieldValue > 0.00) {
-
-                      if (locale.value === 'en') {
-
-                        // ..add missing decimal places and render the amount in the point format.
-                        amountField.value = amountFieldValue.toFixed(2);
-
-                        // If the language is set to German.
-                      } else {
-                        // ..add missing decimal places and render the amount in the comma format.
-                        amountField.value = amountFieldValue.toFixed(2).replace(/\./g, ',');
-                      }
-
-                      // Enable PayPal buttons and remove the warning message.
-                      actions.enable();
-                    } else {
-                      actions.disable();
-                      amountField.setCustomValidity('Invalid field');
-                    }
-                  });
-                },
-                // Set up the transaction
-                createOrder: function (data, actions) {
-                  return actions.order.create({
-                    purchase_units: [{
-                      amount: {
-                        value: formatCurrency(amountField.value)
-                      }
-                    }]
-                  });
-                },
-                onApprove: function (data, actions) {
-                  // gray out and show spinner
-                  return actions.order.capture().then(() => {
-                    return fetch('/payment/ecash/form/submit', {
-                      method: 'post',
-                      headers: {
-                        'content-type': 'application/json'
-                      },
-                      body: JSON.stringify([
-                        { 'name': 'ecash[profile]', 'value': userDataStore.getState().user },
-                        { 'name': 'ecash[orderid]', 'value': data.orderID },
-                        { 'name': 'ecash[amount]', 'value': formatCurrency(amountField.value) }
-                      ])
-                    }).then((response) => {
-                      if(response.ok) {
-                        userDataStore.adjustBalance(parseFloat(formatCurrency(amountField.value)));
-                        transactionStore.fillStore();
-                        // disable gray out and show spinner
-                        vm.$emit('closePanel');
-                      }
-                    });
-                  });
-                },
-              })
-              .render("#paypal-container")
-              .catch((error) => {
-                console.error("failed to render the PayPal Buttons", error);
-              });
-        })
-        .catch((error) => {
-          console.error("failed to load the PayPal JS SDK script", error);
-        })
-        .then(() => this.loaded = true);
+  } catch(error) {
+    console.error('failed to load the PayPal JS SDK script', error);
+    isLoading.value = false;
   }
-}
+  loaded.value = true;
+});
 
+onUnmounted(() => {
+  if (actionsWatcher.value !== null) {
+    actionsWatcher.value();
+  }
+});
+
+function formatCurrency(total: number) {
+  let neg = false;
+  if (total < 0) {
+    neg = true;
+    total = Math.abs(total);
+  }
+  return (neg ? '-' : '') + total.toFixed(2);
+}
 </script>
